requirements.txt
python-dotenv>=0.19.2
requests>=2.26.0
anthropic>=0.2.8
openai>=1.0.0
pillow>=8.4.0
tweepy>=4.4.0
facebook-sdk>=3.1.0
stability-sdk>=0.3.1
opencv-python>=4.5.5.64
numpy>=1.24.0
groq>=0.3.1
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0
transformers>=4.0.0
openai-whisper
# pyaudio - requires portaudio19-dev system package
ollama==0.1.4
beautifulsoup4==4.10.0
flask==2.0.2
werkzeug==2.0.3
elevenlabs==1.7.0
metagen
langchain==0.2.14
langchain-openai==0.1.15
langchain-community==0.2.7
duckduckgo_search==6.1.12
ezsynth
psycopg2-binary
diffusers
matplotlib
accelerate
sentencepiece
huggingface-hub
protobuf
geopy